Add changes derived from 2dust/v2rayN (Can't perform a PR against an unrelated repo). Remove all promotion and update-check informations. Target the default temp path at the system temp dir. Add lazy loaded configuration holder. Display running port for LAN proxy.

+++ v2rayN/Handler/V2rayHandler.cs
@@ -22,6 +22,7 @@ namespace v2rayN.Handler
         private static string v2rayConfigRes = Global.v2rayConfigFileName;
         private List<string> lstV2ray;
         public event ProcessDelegate ProcessEvent;
+        private int processId = 0;
 
         public V2rayHandler()
         {
@@ -39,7 +40,7 @@ namespace v2rayN.Handler
             {
                 string msg = string.Empty;
                 string fileName = Utils.GetPath(v2rayConfigRes);
-                if (V2rayConfigHandler.GenerateClientConfig(config, fileName, out msg) != 0)
+                if (V2rayConfigHandler.GenerateClientConfig(config, fileName, false, out msg) != 0)
                 {
                     ShowMsg(false, msg);
                 }
@@ -67,17 +68,31 @@ namespace v2rayN.Handler
         {
             try
             {
-                foreach (string vName in lstV2ray)
+                bool blExist = true;
+                if (processId > 0)
                 {
-                    Process[] killPro = Process.GetProcessesByName(vName);
-                    foreach (Process p in killPro)
+                    Process p1 = Process.GetProcessById(processId);
+                    if (p1 != null)
                     {
-                        p.Kill();
+                        p1.Kill();
+                        blExist = false;
+                    }
+                }
+                if (blExist)
+                {
+                    foreach (string vName in lstV2ray)
+                    {
+                        Process[] killPro = Process.GetProcessesByName(vName);
+                        foreach (Process p in killPro)
+                        {
+                            p.Kill();
+                        }
                     }
                 }
             }
-            catch (Exception)
+            catch (Exception ex)
             {
+                Utils.SaveLog(ex.Message, ex);
             }
         }
 
@@ -86,7 +101,7 @@ namespace v2rayN.Handler
         /// </summary>
         private void V2rayStart()
         {
-            ShowMsg(false, string.Format("启动服务({0})......", DateTime.Now.ToString()));
+            ShowMsg(false, string.Format(UIRes.I18N("StartService"), DateTime.Now.ToString()));
 
             try
             {
@@ -103,8 +118,8 @@ namespace v2rayN.Handler
                     }
                 }
                 if (Utils.IsNullOrEmpty(fileName))
-                {
-                    string msg = "未找到v2ray文件";
+                {       
+                    string msg = string.Format(UIRes.I18N("NotFoundCore"), @"https://github.com/v2ray/v2ray-core/releases");
                     ShowMsg(true, msg);
                     return;
                 }
@@ -119,16 +134,17 @@ namespace v2rayN.Handler
                     if (!String.IsNullOrEmpty(e.Data))
                     {
                         string msg = e.Data + Environment.NewLine;
-                        //this.AppendText(e.Data + Environment.NewLine);
                         ShowMsg(false, msg);
                     }
                 });
                 p.Start();
                 p.BeginOutputReadLine();
+                processId = p.Id;
             }
-            catch (Exception)
+            catch (Exception ex)
             {
-                string msg = "未找到v2ray文件...";
+                Utils.SaveLog(ex.Message, ex);
+                string msg = ex.Message;
                 ShowMsg(true, msg);
             }
         }
